Address PR review feedback
- Merge duplicate JSDoc blocks in food.ts
- Clarify poop acceleration is an intentional design extension (30-120 ticks)
- Move formatTicksAsTime to common.ts for reusability
- Update items.md spec to document varying poop acceleration values
